+++ geodata/views.py
@@ -1,4 +1,3 @@
-import datetime
 import json
 import os
 
@@ -12,6 +11,7 @@ from django.views.generic import DetailView
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework.decorators import api_view
 
+from rgd import utility
 from . import models
 from . import search
 from .models.imagery.base import RasterEntry
@@ -26,7 +26,16 @@ class RasterEntriesListView(generic.ListView):
         # latitude, longitude, radius, time, timespan, and timefield
         search_params = {}
         time_params = {}
-        for key in {'longitude', 'latitude', 'radius', 'startdate', 'enddate', 'starttime', 'endtime', 'timefield'}:
+        for key in {
+            'longitude',
+            'latitude',
+            'radius',
+            'startdate',
+            'enddate',
+            'starttime',
+            'endtime',
+            'timefield',
+        }:
             if self.request.GET.get(key):
                 try:
                     if key == 'startdate' or key == 'enddate':
@@ -40,10 +49,10 @@ class RasterEntriesListView(generic.ListView):
                 except ValueError:
                     pass
 
-        time, diff = parse_time_params(time_params)
+        time, diff = utility.parse_time_params(time_params)
         if time:
-            search_params['time'] = time
-            search_params['timespan'] = diff
+            search_params['time'] = time.isoformat()
+            search_params['timespan'] = diff.total_seconds()
 
         return self.model.objects.filter(search.search_near_point_filter(search_params))
 
@@ -85,39 +94,3 @@ def download_file(request, model, id, field):
     if len(file) is not None:
         response['Content-Length'] = len(file)
     return response
-
-
-def parse_time_params(time_params):
-    start = None
-    end = None
-    if 'startdate' in time_params:
-        start = time_params['startdate']
-    if 'starttime' in time_params:
-        if start:
-            start = start.replace(hour=time_params['starttime'].hour, minute=time_params['starttime'].minute)
-        else:
-            start = time_params['starttime']
-
-    if 'enddate' in time_params:
-        end = time_params['enddate']
-    if 'endtime' in time_params:
-        if end:
-            end = end.replace(hour=time_params['endtime'].hour, minute=time_params['endtime'].minute)
-        elif start:
-            end = start.replace(hour=time_params['endtime'].hour, minute=time_params['endtime'].minute)
-        else:
-            end = time_params['endtime']
-            
-    if start and end:
-        diff = (end - start) / 2
-        time = (start + diff).isoformat()
-    elif start:
-        next_day = datetime.datetime(start.year, start.month, start.day) + datetime.timedelta(days=1)
-        diff = (next_day - start) / 2
-        time = (start + diff).isoformat()
-    elif end:
-        past_day = datetime.datetime(end.year, end.month, end.day)
-        diff = (end - past_day) / 2
-        time = (end - diff).isoformat()
-
-    return (time, diff.total_seconds())
